[EI-236] [web] refactor: Refactor import path for IntervalToggleGroup in indicator-board.tsx
