+++ web/app/ui/components/numerical-guidance/indicator-board/interval-toggle-group.tsx
